@@ -1,145 +1,133 @@
+<script setup lang="ts">
+import { onBeforeUnmount, onMounted, ref, watch } from 'vue';
+
+import { addClass, removeClass } from '@/utils/index';
+
+const show = ref(false);
+
+defineProps({
+  buttonTop: {
+    default: 250,
+    type: Number
+  }
+});
+
+watch(show, value => {
+  if (value) {
+    addEventClick();
+  }
+  if (value) {
+    addClass(document.body, 'showRightPanel');
+  } else {
+    removeClass(document.body, 'showRightPanel');
+  }
+});
+
+function addEventClick() {
+  window.addEventListener('click', closeSidebar, { passive: true });
+}
+
+function closeSidebar(evt: any) {
+  // 主题选择点击不关闭
+  let parent = evt.target.closest('.right-panel-container');
+  if (!parent) {
+    show.value = false;
+    window.removeEventListener('click', closeSidebar);
+  }
+}
+
+const rightPanel = ref();
+
+function insertToBody() {
+  const body = document.querySelector('body') as any;
+  body.insertBefore(rightPanel.value, body.firstChild);
+}
+
+onMounted(() => {
+  insertToBody();
+});
+
+onBeforeUnmount(() => {
+  rightPanel.value.remove();
+});
+</script>
+
 <template>
-  <div ref="rightPanel" :class="{show:show}" class="rightPanel-container">
-    <div class="rightPanel-background" />
-    <div class="rightPanel">
-      <div class="handle-button" :style="{'top':buttonTop+'px','background-color':theme}" @click="show=!show">
-        <i :class="show?'el-icon-close':'el-icon-setting'" />
+  <div :class="{ show: show }" ref="rightPanel">
+    <div class="right-panel-overlay" />
+    <div class="right-panel-container">
+      <div
+        class="right-panel-btn"
+        :style="{
+          top: buttonTop + 'px'
+        }"
+        @click="show = !show"
+      >
+        <i-ep-close v-show="show" />
+        <i-ep-setting v-show="!show" />
       </div>
-      <div class="rightPanel-items">
+      <div>
         <slot />
       </div>
     </div>
   </div>
 </template>
 
-<script>
-import { addClass, removeClass } from '@/utils'
-
-export default {
-  name: 'RightPanel',
-  props: {
-    clickNotClose: {
-      default: false,
-      type: Boolean
-    },
-    buttonTop: {
-      default: 250,
-      type: Number
-    }
-  },
-  data() {
-    return {
-      show: false
-    }
-  },
-  computed: {
-    theme() {
-      return this.$store.state.settings.theme
-    }
-  },
-  watch: {
-    show(value) {
-      if (value && !this.clickNotClose) {
-        this.addEventClick()
-      }
-      if (value) {
-        addClass(document.body, 'showRightPanel')
-      } else {
-        removeClass(document.body, 'showRightPanel')
-      }
-    }
-  },
-  mounted() {
-    this.insertToBody()
-  },
-  beforeDestroy() {
-    const elx = this.$refs.rightPanel
-    elx.remove()
-  },
-  methods: {
-    addEventClick() {
-      window.addEventListener('click', this.closeSidebar)
-    },
-    closeSidebar(evt) {
-      const parent = evt.target.closest('.rightPanel')
-      if (!parent) {
-        this.show = false
-        window.removeEventListener('click', this.closeSidebar)
-      }
-    },
-    insertToBody() {
-      const elx = this.$refs.rightPanel
-      const body = document.querySelector('body')
-      body.insertBefore(elx, body.firstChild)
-    }
-  }
-}
-</script>
-
-<style>
+<style lang="scss" scoped>
 .showRightPanel {
   overflow: hidden;
   position: relative;
   width: calc(100% - 15px);
 }
-</style>
-
-<style lang="scss" scoped>
-.rightPanel-background {
+.right-panel-overlay {
   position: fixed;
   top: 0;
   left: 0;
-  opacity: 0;
-  transition: opacity .3s cubic-bezier(.7, .3, .1, 1);
-  background: rgba(0, 0, 0, .2);
-  z-index: -1;
+  background: rgba(0, 0, 0, 0.2);
 }
 
-.rightPanel {
+.right-panel-container {
+  background-color: var(--el-bg-color-overlay);
   width: 100%;
-  max-width: 260px;
+  max-width: 300px;
   height: 100vh;
   position: fixed;
   top: 0;
   right: 0;
-  box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, .05);
-  transition: all .25s cubic-bezier(.7, .3, .1, 1);
+  box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.05);
+  transition: all 0.25s cubic-bezier(0.7, 0.3, 0.1, 1);
   transform: translate(100%);
-  background: #fff;
-  z-index: 40000;
+  z-index: 999;
 }
 
 .show {
-  transition: all .3s cubic-bezier(.7, .3, .1, 1);
-
-  .rightPanel-background {
-    z-index: 20000;
+  transition: all 0.3s cubic-bezier(0.7, 0.3, 0.1, 1);
+  .right-panel-overlay {
+    z-index: 99;
     opacity: 1;
     width: 100%;
     height: 100%;
   }
 
-  .rightPanel {
+  .right-panel-container {
     transform: translate(0);
   }
 }
 
-.handle-button {
-  width: 48px;
-  height: 48px;
+.right-panel-btn {
+  background-color: var(--el-color-primary);
+  color: var(--el-color-white);
+  width: 36px;
+  height: 36px;
+  left: -36px;
   position: absolute;
-  left: -48px;
   text-align: center;
-  font-size: 24px;
-  border-radius: 6px 0 0 6px !important;
-  z-index: 0;
-  pointer-events: auto;
+  border-radius: 6px 0 0 6px;
   cursor: pointer;
-  color: #fff;
-  line-height: 48px;
-  i {
-    font-size: 24px;
-    line-height: 48px;
+  svg {
+    vertical-align: -10px;
+    width: 20px;
+    height: 20px;
   }
 }
 </style>
