@@ -1,323 +1,231 @@
 <template>
   <div class="login-container">
-    <el-form ref="loginForm" :model="loginForm" :rules="loginRules" class="login-form" autocomplete="on" label-position="left">
-
-      <div class="title-container">
-        <h3 class="title">Login Form</h3>
+    <el-form
+      ref="loginFormRef"
+      :model="loginData"
+      :rules="loginRules"
+      class="login-form"
+    >
+      <div class="flex text-white items-center py-4">
+        <span class="text-2xl flex-1 text-center">{{ $t('login.title') }}</span>
+        <lang-select style="color: #fff" />
       </div>
 
       <el-form-item prop="username">
-        <span class="svg-container">
+        <div class="p-2 text-white">
           <svg-icon icon-class="user" />
-        </span>
+        </div>
         <el-input
+          class="flex-1"
           ref="username"
-          v-model="loginForm.username"
-          placeholder="Username"
+          size="large"
+          v-model="loginData.username"
+          :placeholder="$t('login.username')"
           name="username"
-          type="text"
-          tabindex="1"
-          autocomplete="on"
         />
       </el-form-item>
 
-      <el-tooltip v-model="capsTooltip" content="Caps lock is On" placement="right" manual>
+      <el-tooltip
+        :disabled="isCapslock === false"
+        content="Caps lock is On"
+        placement="right"
+      >
         <el-form-item prop="password">
-          <span class="svg-container">
+          <span class="p-2 text-white">
             <svg-icon icon-class="password" />
           </span>
           <el-input
-            :key="passwordType"
-            ref="password"
-            v-model="loginForm.password"
-            :type="passwordType"
-            placeholder="Password"
+            class="flex-1"
+            v-model="loginData.password"
+            placeholder="密码"
+            :type="passwordVisible === false ? 'password' : 'input'"
+            size="large"
             name="password"
-            tabindex="2"
-            autocomplete="on"
-            @keyup.native="checkCapslock"
-            @blur="capsTooltip = false"
-            @keyup.enter.native="handleLogin"
+            @keyup="checkCapslock"
+            @keyup.enter="handleLogin"
           />
-          <span class="show-pwd" @click="showPwd">
-            <svg-icon :icon-class="passwordType === 'password' ? 'eye' : 'eye-open'" />
+          <span class="mr-2" @click="passwordVisible = !passwordVisible">
+            <svg-icon
+              :icon-class="passwordVisible === false ? 'eye' : 'eye-open'"
+              class="text-white cursor-pointer"
+            />
           </span>
         </el-form-item>
       </el-tooltip>
 
-      <el-button :loading="loading" type="primary" style="width:100%;margin-bottom:30px;" @click.native.prevent="handleLogin">Login</el-button>
+      <!-- 验证码 -->
+      <el-form-item prop="verifyCode">
+        <span class="p-2 text-white">
+          <svg-icon icon-class="verify_code" />
+        </span>
+        <el-input
+          v-model="loginData.verifyCode"
+          auto-complete="off"
+          :placeholder="$t('login.verifyCode')"
+          class="w-[60%]"
+          @keyup.enter="handleLogin"
+        />
 
-      <div style="position:relative">
-        <div class="tips">
-          <span>Username : admin</span>
-          <span>Password : any</span>
+        <div class="captcha">
+          <img :src="captchaBase64" @click="getCaptcha" />
         </div>
-        <div class="tips">
-          <span style="margin-right:18px;">Username : editor</span>
-          <span>Password : any</span>
-        </div>
+      </el-form-item>
 
-        <el-button class="thirdparty-button" type="primary" @click="showDialog=true">
-          Or connect with
-        </el-button>
+      <el-button
+        size="default"
+        :loading="loading"
+        type="primary"
+        class="w-full"
+        @click.prevent="handleLogin"
+        >{{ $t('login.login') }}
+      </el-button>
+
+      <!-- 账号密码提示 -->
+      <div class="mt-4 text-white text-sm">
+        <span>{{ $t('login.username') }}: admin</span>
+        <span class="ml-4"> {{ $t('login.password') }}: 123456</span>
       </div>
     </el-form>
-
-    <el-dialog title="Or connect with" :visible.sync="showDialog">
-      Can not be simulated on local, so please combine you own business simulation! ! !
-      <br>
-      <br>
-      <br>
-      <social-sign />
-    </el-dialog>
   </div>
 </template>
 
-<script>
-import { validUsername } from '@/utils/validate'
-import SocialSign from './components/SocialSignin'
+<script setup lang="ts">
+import router from '@/router';
+import LangSelect from '@/components/LangSelect/index.vue';
+import SvgIcon from '@/components/SvgIcon/index.vue';
 
-export default {
-  name: 'Login',
-  components: { SocialSign },
-  data() {
-    const validateUsername = (rule, value, callback) => {
-      if (!validUsername(value)) {
-        callback(new Error('Please enter the correct user name'))
-      } else {
-        callback()
-      }
-    }
-    const validatePassword = (rule, value, callback) => {
-      if (value.length < 6) {
-        callback(new Error('The password can not be less than 6 digits'))
-      } else {
-        callback()
-      }
-    }
-    return {
-      loginForm: {
-        username: 'admin',
-        password: '111111'
-      },
-      loginRules: {
-        username: [{ required: true, trigger: 'blur', validator: validateUsername }],
-        password: [{ required: true, trigger: 'blur', validator: validatePassword }]
-      },
-      passwordType: 'password',
-      capsTooltip: false,
-      loading: false,
-      showDialog: false,
-      redirect: undefined,
-      otherQuery: {}
-    }
-  },
-  watch: {
-    $route: {
-      handler: function(route) {
-        const query = route.query
-        if (query) {
-          this.redirect = query.redirect
-          this.otherQuery = this.getOtherQuery(query)
-        }
-      },
-      immediate: true
-    }
-  },
-  created() {
-    // window.addEventListener('storage', this.afterQRScan)
-  },
-  mounted() {
-    if (this.loginForm.username === '') {
-      this.$refs.username.focus()
-    } else if (this.loginForm.password === '') {
-      this.$refs.password.focus()
-    }
-  },
-  destroyed() {
-    // window.removeEventListener('storage', this.afterQRScan)
-  },
-  methods: {
-    checkCapslock(e) {
-      const { key } = e
-      this.capsTooltip = key && key.length === 1 && (key >= 'A' && key <= 'Z')
-    },
-    showPwd() {
-      if (this.passwordType === 'password') {
-        this.passwordType = ''
-      } else {
-        this.passwordType = 'password'
-      }
-      this.$nextTick(() => {
-        this.$refs.password.focus()
-      })
-    },
-    handleLogin() {
-      this.$refs.loginForm.validate(valid => {
-        if (valid) {
-          this.loading = true
-          this.$store.dispatch('user/login', this.loginForm)
-            .then(() => {
-              this.$router.push({ path: this.redirect || '/', query: this.otherQuery })
-              this.loading = false
-            })
-            .catch(() => {
-              this.loading = false
-            })
-        } else {
-          console.log('error submit!!')
-          return false
-        }
-      })
-    },
-    getOtherQuery(query) {
-      return Object.keys(query).reduce((acc, cur) => {
-        if (cur !== 'redirect') {
-          acc[cur] = query[cur]
-        }
-        return acc
-      }, {})
-    }
-    // afterQRScan() {
-    //   if (e.key === 'x-admin-oauth-code') {
-    //     const code = getQueryObject(e.newValue)
-    //     const codeMap = {
-    //       wechat: 'code',
-    //       tencent: 'code'
-    //     }
-    //     const type = codeMap[this.auth_type]
-    //     const codeName = code[type]
-    //     if (codeName) {
-    //       this.$store.dispatch('LoginByThirdparty', codeName).then(() => {
-    //         this.$router.push({ path: this.redirect || '/' })
-    //       })
-    //     } else {
-    //       alert('第三方登录失败')
-    //     }
-    //   }
-    // }
+// 状态管理依赖
+import { useUserStore } from '@/store/modules/user';
+
+// API依赖
+import { LocationQuery, LocationQueryValue, useRoute } from 'vue-router';
+
+const userStore = useUserStore();
+const route = useRoute();
+
+/**
+ * 按钮loading
+ */
+const loading = ref(false);
+/**
+ * 是否大写锁定
+ */
+const isCapslock = ref(false);
+/**
+ * 密码是否可见
+ */
+const passwordVisible = ref(false);
+/**
+ * 验证码图片Base64字符串
+ */
+const captchaBase64 = ref();
+
+/**
+ * 登录表单引用
+ */
+const loginFormRef = ref(ElForm);
+
+
+const loginRules = {
+  username: [{ required: true, trigger: 'blur' }],
+  password: [{ required: true, trigger: 'blur', validator: passwordValidator }],
+  verifyCode: [{ required: true, trigger: 'blur' }]
+};
+
+/**
+ * 密码校验器
+ */
+function passwordValidator(rule: any, value: any, callback: any) {
+  if (value.length < 6) {
+    callback(new Error('The password can not be less than 6 digits'));
+  } else {
+    callback();
   }
 }
+
+/**
+ * 检查输入大小写状态
+ */
+function checkCapslock(e: any) {
+  const { key } = e;
+  isCapslock.value = key && key.length === 1 && key >= 'A' && key <= 'Z';
+}
+
+/**
+ * 登录
+ */
+function handleLogin() {
+
+}
+
+onMounted(() => {
+});
 </script>
 
-<style lang="scss">
-/* 修复input 背景不协调 和光标变色 */
-/* Detail see https://github.com/PanJiaChen/vue-element-admin/pull/927 */
-
-$bg:#283443;
-$light_gray:#fff;
-$cursor: #fff;
-
-@supports (-webkit-mask: none) and (not (cater-color: $cursor)) {
-  .login-container .el-input input {
-    color: $cursor;
-  }
-}
-
-/* reset element-ui css */
-.login-container {
-  .el-input {
-    display: inline-block;
-    height: 47px;
-    width: 85%;
-
-    input {
-      background: transparent;
-      border: 0px;
-      -webkit-appearance: none;
-      border-radius: 0px;
-      padding: 12px 5px 12px 15px;
-      color: $light_gray;
-      height: 47px;
-      caret-color: $cursor;
-
-      &:-webkit-autofill {
-        box-shadow: 0 0 0px 1000px $bg inset !important;
-        -webkit-text-fill-color: $cursor !important;
-      }
-    }
-  }
-
-  .el-form-item {
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    background: rgba(0, 0, 0, 0.1);
-    border-radius: 5px;
-    color: #454545;
-  }
-}
-</style>
-
 <style lang="scss" scoped>
-$bg:#2d3a4b;
-$dark_gray:#889aa4;
-$light_gray:#eee;
-
 .login-container {
   min-height: 100%;
   width: 100%;
-  background-color: $bg;
+  background-color: #2d3a4b;
   overflow: hidden;
 
   .login-form {
-    position: relative;
     width: 520px;
     max-width: 100%;
     padding: 160px 35px 0;
     margin: 0 auto;
     overflow: hidden;
-  }
 
-  .tips {
-    font-size: 14px;
-    color: #fff;
-    margin-bottom: 10px;
+    .captcha {
+      position: absolute;
+      right: 0;
+      top: 0;
 
-    span {
-      &:first-of-type {
-        margin-right: 16px;
+      img {
+        height: 48px;
+        width: 120px;
+        cursor: pointer;
       }
     }
   }
+}
 
-  .svg-container {
-    padding: 6px 5px 6px 15px;
-    color: $dark_gray;
-    vertical-align: middle;
-    width: 30px;
-    display: inline-block;
-  }
+.el-form-item {
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  background: rgba(0, 0, 0, 0.1);
+  border-radius: 5px;
+}
 
-  .title-container {
-    position: relative;
+.el-input {
+  background: transparent;
 
-    .title {
-      font-size: 26px;
-      color: $light_gray;
-      margin: 0px auto 40px auto;
-      text-align: center;
-      font-weight: bold;
-    }
-  }
+  // 子组件 scoped 无效，使用 :deep
+  :deep(.el-input__wrapper) {
+    padding: 0;
+    background: transparent;
+    box-shadow: none;
 
-  .show-pwd {
-    position: absolute;
-    right: 10px;
-    top: 7px;
-    font-size: 16px;
-    color: $dark_gray;
-    cursor: pointer;
-    user-select: none;
-  }
+    .el-input__inner {
+      background: transparent;
+      border: 0px;
+      border-radius: 0px;
+      color: #fff;
+      caret-color: #fff;
 
-  .thirdparty-button {
-    position: absolute;
-    right: 0;
-    bottom: 6px;
-  }
+      &:-webkit-autofill {
+        box-shadow: 0 0 0 1000px transparent inset !important;
+        -webkit-text-fill-color: #fff !important;
+      }
 
-  @media only screen and (max-width: 470px) {
-    .thirdparty-button {
-      display: none;
+      // 设置输入框自动填充的延迟属性
+      &:-webkit-autofill,
+      &:-webkit-autofill:hover,
+      &:-webkit-autofill:focus,
+      &:-webkit-autofill:active {
+        -webkit-transition-delay: 99999s;
+        -webkit-transition: color 99999s ease-out,
+          background-color 99999s ease-out;
+      }
     }
   }
 }
